+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Added
 - Are you sure message when closing Pext normally
+- Add installable quick action service for easier launching
 
 ### Changed
 - Pick a more reasonable height on wide screens